+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Fix "Break" box gui not displaying on client instances and remote
 - Fix combatant status effects not decaying
 - Fix food icon not dying on remote when eating is interrupted
+- Fix attack circle glow only working for the first remote player on map
 
 ## [0.8.2] 2026-07-26